Fix comment to match code
But, as the comment says, there is something weird going on there, and
I might well have misunderstood something completely last week. I was
confused then as the wrong function was called because I had wrongly
skipped generating one virtual function from an interface, and the
parameters seemed to be in the wrong order for that incorrect
function, which I thought was somehow automatically called by the
VB6-generated client, when what it was calling was in fact the right
one, but the vtable of the generated proxy class was incorrect.

Change-Id: I1661d746459f80076a45c401b2b7d286bda7f09d