Add a program to output the threadProc() code as an initialized C array
(But I think this won't actually be needed. Read on.)

Otherwise we can't have a single exewrapper executable that handles
both 32- and 64-bit wrapped executables, as obviously a function
included in exewrapper itself will be compiled just for the
architecture exewrapper is compiled for.

But do we really want that anyway? Actually having the one same
exewrapper executable (built as 32- or 64-bit) inject code into
another process that might be of a different bitness will be quite
hard. Not because of injecting the correct code as such (that is what
this program in this commit takes care of), but passing the addresses
of LoadLibraryW etc in the struct to the remote thread
procedure. Surely the address of LoadLibraryW etc is not the same in
32- and 64-bit processes.

Change-Id: I405ade70e9fb24d3981db7f855943961b9a51f94