Start work on handling both 32- and 64-bit wrapped processes
Introduce a higher level of wrapper process. This is a program called
"coleat". It is this program that creates the process of the program
to be wrapped (with its tsratup thread suspended, as before). The
coleat program can be built as 32- or 64-bit, it doesn't matter. Also
the wrapped program can be either.

Once the wrapped program process is created, coleat can find out
whether it is 32- or 64-bit. It then starts the exewrapper
program. The exewrapper program should be available in two flavours,
32- and 64-bit. (They are distinguished by having -x86 or -x64
appended to their basename.)  Coleat starts the one that matches the
wrapped procgram. Coleat passes it inheritable handles to the wrapped
program's process and startup thread.

Exewrapper then works like it did before, the only difference is that
now it doesn't create the processs of the wrapped rogram, but uses the
handle for it that it got from coleat.

Note: The function hooking is most likely broken for 64-bit. It hasn't
been tested at al, might not even compile. Will look into that
tomorrow.

Change-Id: Ia8627b76653ecf4fce5e7e8e0af902b5b8db6da7