Give the developer a chance to attach a debugger also to the wrapped process
Don't use the -d switch lightly. It really does require you to be
debugging it actively, and manually in the debugger change the value
of a variable, so that the code gets out of an infinite loop.

Mention the name of the exe file that should be attached, to make it
easier to find them in VS's 'Attach to Process' dialog.

Change-Id: Idb439eadff0ed58261911072863277e835d5779a