Drop the -m and -M options from the coleat command line
Instead, just have genproxy take such an argument instead, and compile
in the thusly specified mappings into the DLL to be injected. We
compile in a fixed set of generated proxies for specific interfaces,
too, so no harm done in having the corresponding coclass mappings also
built-in.

Rename genproxy's -M option to be -O, and -M is now the option to
specify the coclass interface map.

Change-Id: I505ead48f78d3d24558d051142880967088a6cae