What I called "IID2" or "Extra IID" is simply the coclass of the interface
So rename accordingly.

But actually, I have a feeling that I need to re-think and re-work
much of how the proxy classes hang together. Some even simple things
break in the (in progress) tracing-only mode. There must be something
fundamentally wrong. See the rules for implementing IUnknown on
https://msdn.microsoft.com/en-us/library/ms810016.aspx . I should try
harder to make sure they are followed by the proxy objects.

Change-Id: I3fb991a7f35984982524bc1097cc8d32cc8b9379