Drop trailing empty VARIANT parameters before calling genericInvoke()
This was the problem that I had with tracing-only mode of an Excel
client, sigh. I wasted much time chasing wrong ideas before I finally
noticed that the Invoke of the Open methiod actually returned an
error.

Update the TODO and bin temporary debugging output added while chasing
the problem.

Change-Id: Ief8b8eaa5a538e2b40c23b206c1e513284896d03