Can simplify CProxiedSink::Invoke() muchly in the tracing-only case
We actually don't seem to need the name:memberid mapping from
OutgoingInterfaceMap.hxx after all.

In the tracing-only case, the traced client program connects to the
same "original" application that we generated the proxies and mapping
data from. (From its type library.)

That is the same type information that the client program has used
when creating its sink interfaces. When the application does a
callback, that our layer catches, the incoming memberid is thus the
same that should be used when we complete the callback down to the
program being traced.

Change-Id: Idac2e0b59dbf8fbe073185b0deb7885e4836e54b