Start splitting output into "verbose" and "trace" output
The plan is that with -t (trace only), the output should look a bit
like pseudo-code representing what APIs the wrapped client calls in
the application, what parameters are passed, and return values.

The -v (verbose) output includes all kinds of details about the proxy
classes and their invocations.

With neither option, coleat is silent, except for error messages.