Proxy also return values from functions invoked from a late binding client
When wrapping compile-time-bound clients all function calls in proxied
classes to thouth the generated proxy function that creates a proxy
object for a returned object that is of a type that interests us.

Now do the same when wrapping late-binding clients, where it is
CProxiedDispatch::Invoke() that does it. Also add printing of the
return value to it. (We still don't print parameters there.)

For this to work, we need to generate yet another file,
ProxyCreator.hxx.

While at it, add some more utility functions to util.hpp.