Re-work error and verbose reporting from injecteddll.cpp
While running the injected thread, use an event in the parameter block
to signal to the exewrapper process that there is a message to perhaps
display. Once the injected thread has finished, the code in
injecteddll.cpp can use the std streams normally.

Add a parameter to hook() and storeMessage() to indicate whether the
function to hook is mnadatory to be present.

Rename storeError() to storeMessage() as it isn't just for errors now.

Clean up the silly uses of several sequential wcscpy_s() and
wcscat_s() calls to store a message; just add a few more optional
parameters to storeMessage() instead so that we can use that always.