Hook DllGetClassObject() functions in DLLs using trampolines
As there can be a DllGetClassObject() function in a number of DLLs
used during the client app runtime, we must be able to find out in our
hooked DllGetClassObject() which DLL it is that we are handling, ir
order to be able to call the real one first.

The plan is that if a DllGetClassObject() is for a type we know (have
generated proxy classes for), we can proxy the return value. (And if
we are brave and trust that only late-binding calls are done on it,
even if we don't know the type.) This is not done yet.