Do not use stderr in the run-time code
(It is fine to use it in genproxy which is run by developers.)

Making sure just stdout works on Windows 7 and older is hard
enough. No need to make life any harder by introducing additional
complexity with both stdout and stderr, like if one of them is OK but
not the other, and what to do then.

This commit doesn't fix the actual problem we see on Windows 7, where
if not redirected, only the first few lines of output are displayed in
the console window. It just makes an eventual fix simpler when stderr
(GetStdHandle(STD_ERROR_HANDLE), STARTUPINFO::hStdError, std::cerr)
can be ignored.