Progress towards making this work better as a GUI tool
Create a copy of coleat.exe called coleatw.exe that is edited with
link -edit to be -subsystem:windows and thus not automatically create
a console window when run.

Interactive cmd.exe users run coleat.exe, otherwise it is coleatw.exe
that one should run (and typically with output redirected to a file).

If coleatw is run without output being redirected to a file, and there
is no parent console to use, open a new console windows and output
there.

Add a possibility to expand a timestamp in the output file name, so
that there can be a shortcut that starts coleatw on some program and
each time creates a different log file. The sequence "%t" is expanded
to a timestamp in the format "20180513.222540". This functionality
could be expanded with other sequences, like for the basename of the
wrapped propgram.