Handle enum types as parameters to generated proxy method better
We should hardly pass them as a VT_USERDEFINED in the DISPPARAMS but
as the actual underlying integral type.

Various other fixes and changes as fallout. Fix syntax error in
generated enum header: Add missing colon after the enum definition.

Add a call to close() to the OutputFile destructor. Makes the code
simpler, no need to explicitly call close() on them.