Actually erase the right map entry
Likely fixes various crashes and assertion failures I have seen. The
code tried to erase a nonexistent map entry, which must have lead to
stale map entries for dead objects in the map, and if some other
object to be proxied later was created at the same address, the old,
bogus, proxy object was re-used for it, with random results.