Re-think how ProxyCreator() works
Instead of taking both an object and the IID of its type, only take an
object. Then in the code call QueryInterface() on it for each of the
interface types we have proxies for. If one (and no more) matches,
create a corresponding proxy.

Also, have ProxyCreator() return the "pretty" name of the interface
for which a proxy was created, if any.

This simplifies the code in CProxiedDispatch::Invoke() significantly.
No more need to try to retrieve type information to get the IID of a
returned object. (Type information might not be available.)

Use ProxyCreator() also CProxiedEnumVARIANT::Next().

Use the pretty type name returned from ProxyCreator() when logging
returned IDIspatch pointers. For instance when enumerating paragraphs,
instead of:

    Enum<006E3E70>.Next(1): S_OK
    ... 0: 00697C6C

we now print:

    Enum<00855920>.Next(1): S_OK
    ... 0: Word.Paragraph<008D1620>

As the values returned when enumerating now are proxied, too, we can
also trace calls to their API. For instance on the above:

    Word.Paragraph<008D1620>.Range -> Word.Range<00673890>