Add Excel.Sheets interface to the sample list
I probably will need to modify the logic in genproxy so that it won't
traverse the type library from one starting point, but instead just go
through it interface by interface and handle those that match our
list. As genproxy currently works it will miss interfaces not
referenced directly by the interfaces we select to handle, even if
mentioned in the list. For instance Excel.Worksheet won't be handled
even if I include it in the list.