Re-work genproxy: Unless asked to, don't traverse starting at a root type
Instead, handle all interfaces requested by the -i option. This way we
will include also those not specifically referenced to by methods or
properties in types that we handle. For instance those returned from
IEnumVARIANT style enumerators as plain IDispatch pointers.

This also means that now we will actually generate a proxy for the
Word.Footnote interface that we have mentioned in mso.interfaces.list
for a while already.