Add special-purpose hack to provide an OleCreateLink() preview
This is specific for a Collabora Productivity customer case and behind
the #ifdef HARDCODE_MSO_TO_CO.

When our OleCreateLink() hook is asked to create an OLERENDER_DRAW
link for a document in a file of a type that Collabora Office handles,
we instead create a link to a dummy empty RTF file. We also run
Collabora Office to produce a PNG image of the first page of the
document, using the --convert-to functionality.

The link to the dummy empty RTF file will be handled by code that
always is present in Windows.

We wrap the returned IOleObject with an own proxy object that
otherwise works identically, except that when QueryInterface() is
called for IID_IViewObject, it returns another wrapper. That wrapper
again otherwise works identically to the IViewObject that the wrapped
IOleObject returned, exept that its Draw() method actually draws the
PNG that Collabora Office produced.