Use a throwaway UserInstallation for the OleCreateLink functionality
We don't want the soffice we run for the --convert-to to interfere
with a possibly already running Collabora Office. Besides, even if
there would be no visible interference, the problem is that if a
soffice.bin process already exists, running soffice.exe --convert-to
will not wait for that soffice.bin to do the --convert-to, so there is
no way to know when it is done.