Calling the myViewObject's advise sink in Draw() was not a good idea, it seems
It caused a loop in the customer application, it kept calling Draw()
again and again... The intended use of that sink must be something
else. (Which kinda makes sense, because presumably it is the
application code (or the VB6 machinery) that causes Draw() to be
called, so why would it need to be notified that Draw() has been
called? Probably this is supposed to work the other way, a call to the
advice sink's OnViewChange() will cause Draw() to be called?)